dataset/requirements.txt
pillow
pandas
numpy
megadetector
ipykernel
minio
matplotlib
seaborn